+++ src/reactive/Field.js
@@ -7,8 +7,8 @@ import React, {PropTypes} from 'react';
 import * as Stylesheet from 'react-stylesheet';
 import Component from '../Component';
 import Input from '../Input';
-import ErrorList from '../ErrorList';
 import Label from '../Label';
+import ErrorList from './ErrorList';
 import reactive from './reactive';
 
 class FieldBase extends Component {
