+++ src/reactive/index.js
@@ -7,7 +7,8 @@ export * as Schema from '../Schema';
 export Fieldset from '../Fieldset';
 export Input from '../Input';
 
-export withFormValue from '../withFormValue';
+export withFormValue from './withFormValue';
+export {create as createValue} from './Value';
 export reactive from './reactive';
 export ErrorList from './ErrorList';
 export Field from './Field';
